+++ CHANGELOG.md
@@ -10,6 +10,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - Subscriptions count for authenticated dao top request
 
+### Fixed
+- Fixed avatars links in dao/top endpoint
+
 ## [0.0.7] - 2023-07-12
 
 ### Added